Handle continuation of large sysexes split by ALSA
ALSA's rawmidi "virtual" plugin enforces a hardcoded limit of 256 bytes
per generated sequencer event.  The PS/1 version of Silpheed includes a
MT32.DRV which sends a SysEx message that is 264 bytes long, which is
broken into two consecutive incomplete SND_SEQ_EVENT_SYSEX sequencer
event segments.  We have to reassemble the segments for the MT-32 to be
initialized correctly.

Before this workaround, a user reported incorrect instruments with this
particular version of Silpheed, when using mt32emu-qt through DOSEMU's
ALSA rawmidi interface.

This workaround should also be applicable to the old mt32_alsadrv as it
uses the same interface.